feat(edit-content) fix comments
