+++ core-web/libs/edit-content/src/lib/components/dot-edit-content-sidebar/components/dot-edit-content-sidebar-information/dot-edit-content-sidebar-information.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { ChangeDetectionStrategy, Component, computed, input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, computed, inject, input } from '@angular/core';
 import { RouterLink } from '@angular/router';
 
 import { ChipModule } from 'primeng/chip';
@@ -7,13 +7,9 @@ import { SkeletonModule } from 'primeng/skeleton';
 import { TooltipModule } from 'primeng/tooltip';
 
 import { DotCMSContentlet, DotCMSContentType } from '@dotcms/dotcms-models';
-import {
-    DotCopyButtonComponent,
-    DotLinkComponent,
-    DotMessagePipe,
-    DotRelativeDatePipe
-} from '@dotcms/ui';
+import { DotMessagePipe, DotRelativeDatePipe } from '@dotcms/ui';
 
+import { DotMessageService } from '../../../../../../../data-access/src/lib/dot-messages/dot-messages.service';
 import { ContentletStatusPipe } from '../../../../pipes/contentlet-status.pipe';
 import { DotNameFormatPipe } from '../../../../pipes/name-format.pipe';
 
@@ -36,8 +32,6 @@ interface ContentSidebarInformation {
         TooltipModule,
         ChipModule,
         SkeletonModule,
-        DotCopyButtonComponent,
-        DotLinkComponent,
         ContentletStatusPipe,
         DotRelativeDatePipe,
         DotMessagePipe,
@@ -48,6 +42,7 @@ interface ContentSidebarInformation {
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DotEditContentSidebarInformationComponent {
+    #dotMessageService = inject(DotMessageService);
     /**
      * Input that contains the data of the contentlet.
      */
@@ -57,4 +52,14 @@ export class DotEditContentSidebarInformationComponent {
      * Computed that contains the url to the contentlet.
      */
     $jsonUrl = computed(() => `/api/content/id/${this.$data().contentlet.identifier}`);
+
+    /**
+     * Computed that returns a tooltip message when creation date doesn't exist
+     */
+    $createdTooltipMessage = computed(() => {
+        const { contentlet } = this.$data();
+        return !contentlet?.creationDate
+            ? this.#dotMessageService.get('edit.content.sidebar.information.no.created.yet')
+            : null;
+    });
 }
