+++ core-web/libs/edit-content/src/lib/components/dot-edit-content-sidebar/components/dot-edit-content-sidebar-workflow/dot-edit-content-sidebar-workflow.component.html
@@ -20,7 +20,7 @@
                 data-testId="select-workflow-link"
                 (click)="$event.preventDefault(); showDialog()"
                 tabindex="0">
-                {{ 'Select Workflow' | dm }}
+                {{ 'edit.content.sidebar.workflow.select.workflow' | dm }}
             </a>
         } @else {
             {{ scheme?.name }}
