+++ core-web/libs/edit-content/src/lib/components/dot-edit-content-sidebar/components/dot-edit-content-sidebar-workflow/dot-edit-content-sidebar-workflow.component.scss
@@ -12,7 +12,9 @@ $skeleton-height: 14px;
     dt {
         font-weight: $font-weight-medium-bold;
         color: $black;
-        min-height: 32px;
+
+        
+        min-height: 2.286rem;
         display: flex;
         align-content: center;
         flex-wrap: wrap;
