+++ core-web/libs/edit-content/src/lib/components/dot-edit-content-sidebar/components/dot-edit-content-sidebar-workflow/dot-edit-content-sidebar-workflow.component.ts
@@ -66,7 +66,7 @@ export class DotEditContentSidebarWorkflowComponent {
      * @type {string}
      * @memberof DotEditContentSidebarWorkflowComponent
      */
-    $selectedWorkflow = model<SelectItem>({} as SelectItem);
+    $selectedWorkflow = model<SelectItem | null>();
 
     /**
      * The workflow status of the content item.
