+++ core-web/libs/edit-content/src/lib/utils/functions.util.ts
@@ -3,13 +3,9 @@ import {
     DotCMSContentTypeField,
     DotCMSContentTypeFieldVariable,
     DotCMSContentTypeLayoutRow,
-    DotCMSContentTypeLayoutTab,
-    DotCMSWorkflow,
-    DotCMSWorkflowAction,
-    WorkflowStep
+    DotCMSContentTypeLayoutTab
 } from '@dotcms/dotcms-models';
 
-import { ContentState } from '../feature/edit-content/store/features/content.feature';
 import {
     CALENDAR_FIELD_TYPES,
     FLATTENED_FIELD_TYPES,
@@ -297,137 +293,3 @@ export const transformFormDataFn = (contentType: DotCMSContentType): Tab[] => {
         }))
     }));
 };
-
-/**
- * Parses an array of workflow data and returns a new object with key-value pairs.
- *
- * @param {Object[]} data - The array of workflow data to be parsed.
- * @returns {Object} - The parsed object with key-value pairs.
- */
-export const parseWorkflows = (
-    data: {
-        scheme: DotCMSWorkflow;
-        action: DotCMSWorkflowAction;
-        firstStep: WorkflowStep;
-    }[]
-) => {
-    if (!Array.isArray(data)) {
-        return {};
-    }
-
-    return data.reduce((acc, { scheme, action, firstStep }) => {
-        if (!acc[scheme.id]) {
-            acc[scheme.id] = {
-                scheme: {
-                    ...scheme
-                },
-                actions: [],
-                firstStep
-            };
-        }
-
-        acc[scheme.id].actions.push(action);
-
-        return acc;
-    }, {});
-};
-
-/**
- * Determines if workflow action buttons should be shown based on content and scheme state
- * Shows workflow buttons when:
- * - Content type has only one workflow scheme OR
- * - Content is existing AND has a selected workflow scheme OR
- * - Content is new and has selected a workflow scheme
- *
- * @param schemes - Available workflow schemes object
- * @param contentlet - Current contentlet (if exists)
- * @param currentSchemeId - Selected workflow scheme ID
- * @returns boolean indicating if workflow actions should be shown
- */
-export function shouldShowWorkflowActions(
-    schemes: ContentState['schemes'],
-    contentlet: ContentState['contentlet'],
-    currentSchemeId: string | null
-): boolean {
-    const hasOneScheme = Object.keys(schemes).length === 1;
-    const isExisting = !!contentlet;
-    const hasSelectedScheme = !!currentSchemeId;
-
-    if (hasOneScheme) {
-        return true;
-    }
-
-    if (isExisting && hasSelectedScheme) {
-        return true;
-    }
-
-    if (!isExisting && hasSelectedScheme) {
-        return true;
-    }
-
-    return false;
-}
-
-/**
- * Determines if workflow selection warning should be shown
- * Shows warning when:
- * - Content is new (no contentlet exists) AND
- * - Content type has multiple workflow schemes AND
- * - No workflow scheme has been selected
- *
- * @param schemes - Available workflow schemes object
- * @param contentlet - Current contentlet (if exists)
- * @param currentSchemeId - Selected workflow scheme ID
- * @returns boolean indicating if workflow selection warning should be shown
- */
-export function shouldShowWorkflowWarning(
-    schemes: ContentState['schemes'],
-    contentlet: ContentState['contentlet'],
-    currentSchemeId: string | null
-): boolean {
-    const isNew = !contentlet;
-    const hasNoSchemeSelected = !currentSchemeId;
-    const hasMultipleSchemas = Object.keys(schemes).length > 1;
-
-    return isNew && hasMultipleSchemas && hasNoSchemeSelected;
-}
-
-/**
- * Gets the appropriate workflow actions based on content state
- * Returns:
- * - Empty array if no scheme is selected
- * - Current content actions for existing content
- * - Sorted scheme actions for new content (with 'Save' action first)
- *
- * @param schemes - Available workflow schemes object
- * @param contentlet - Current contentlet (if exists)
- * @param currentSchemeId - Selected workflow scheme ID
- * @param currentContentActions - Current content specific actions
- * @returns Array of workflow actions
- */
-export function getWorkflowActions(
-    schemes: ContentState['schemes'],
-    contentlet: ContentState['contentlet'],
-    currentSchemeId: string | null,
-    currentContentActions: DotCMSWorkflowAction[]
-): DotCMSWorkflowAction[] {
-    const isNew = !contentlet;
-
-    // If no scheme is selected, return empty array
-    if (!currentSchemeId || !schemes[currentSchemeId]) {
-        return [];
-    }
-
-    // For existing content, use specific contentlet actions
-    if (!isNew && currentContentActions.length) {
-        return currentContentActions;
-    }
-
-    // For new content, use scheme actions sorted with 'Save' first
-    return Object.values(schemes[currentSchemeId].actions).sort((a, b) => {
-        if (a.name === 'Save') return -1;
-        if (b.name === 'Save') return 1;
-
-        return a.name.localeCompare(b.name);
-    });
-}
