+++ dotCMS/src/main/webapp/WEB-INF/messages/Language.properties
@@ -5840,8 +5840,19 @@ edit.content.layout.select.workflow.warning.switch=Select a Workflow
 edit.content.layout.select.workflow.warning.subtitle=to take action on this content.
 edit.content.sidebar.workflow.dialog.title=Select the workflow you want to work on.
 edit.content.sidebar.workflow.dialog.dropdown.placeholder=Select a Workflow
+edit.content.sidebar.workflow.select.workflow=Select Workflow
+edit.content.sidebar.workflow.you=You
+
+
+edit.content.sidebar.information.moddate.tooltip=Not modified yet
+edit.content.sidebar.information.publishdate.tooltip=Not published yet
+edit.content.sidebar.information.select.workflow=Select the workflow you want to work on
+edit.content.sidebar.information.error.initializing.content=Error initializing content
+edit.content.sidebar.information.go.to.content.type=Go to edit Content Type
+edit.content.sidebar.information.no.created.yet=Not created yet
 
 edit.content.sidebar.information.no.workflow.selected.yet=Select a Workflow to see its data.
+
 
 
 lts.expired.message = This version of dotCMS is no longer supported. Please contact your customer success manager to schedule an upgrade.
